requirements.txt
google-cloud-aiplatform==1.36.4
langchain==0.0.336
openai==1.3.2
sentence-transformers==2.2.2
langchain[docarray]==0.0.336
docarray==0.39.1
pydantic==1.10.13
flask~=3.0.0